implement gunicorn and nginx configuration
