+++ audio_model/audio_model/preprocessing/mp3_parser.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 from audio_model.audio_model.config import config
-from audio_model.audio_model.utils import audio_melspectrogram, audio_mfcc
+from audio_model.audio_model.utils import audio_mfcc
 
 warnings.filterwarnings("ignore")
