--- commonvoice/tox.ini
@@ -1,18 +1,0 @@
-[tox]
-envlist = py37
-skipsdist = True
-
-[testenv]
-install_command =
-	pip install {opts} {packages}
-
-deps =
-	-rrequirements.txt
-
-setenv =
-  	PYTHONPATH= .
-
-commands:
-	pip install torch===1.5.1 torchvision===0.6.1 -f https://download.pytorch.org/whl/torch_stable.html
-
-
