+++ run.sh
@@ -1,2 +1,2 @@
 export IS_DEBUG=${DEBUG:-false}
-exec gunicorn --bind "0.0.0.0:${PORT:-8080}" --access-logfile - --error-logfile - run_app:app
+exec gunicorn --bind "0.0.0.0:${PORT:-8080}" --workers=4 --access-logfile - --error-logfile - run_app:app
