+++ server/nginx.conf
@@ -1,0 +1,9 @@
+server {
+
+    listen 80;
+    server_name commmvoice.me www.commmvoice.me;
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass flask:8080;
+    }
