+++ server/nginx/nginx.conf
@@ -1,0 +1,12 @@
+user www-data;
+...
+http {
+    server {
+        listen          8080;
+        server_name     127.0.0.1;
+        location / {
+            proxy_pass http://unix:/run/gunicorn.sock;
+        }
+    }
+}
+...
